Apply incremental stat scaling using EnemyWaveConfig resource and the associated factories.

+++ start-blastin/architecture/interfaces/IHealthful.cs
@@ -5,7 +5,7 @@ namespace Interfaces
 {
     public interface IHealthful
     {
-        void TakeDamage(int damage);
-        void Heal(int healAmount);
+        void TakeDamage(float damage);
+        void Heal(float healAmount);
     }
 }
